refacotr: 로또 결과 계산하는 함수 디렉토리 변경
- 상태 데이터를 보유하고 있지 않는 lottoMatchResult를 model 클래스에서 util로 변경
